% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_fit.R
\name{linear_fit}
\alias{linear_fit}
\title{linear model fitting function}
\usage{
linear_fit(y, X, n, ndraw = 2000, nomit = 1000)
}
\arguments{
\item{y}{y is the response;}

\item{X}{X is data frame ccontains data;}

\item{n}{n is the observation number;}

\item{ndraw}{ndraw is the draw times for MCMC;}

\item{nomit}{noimt is the draw times omit for burn in;}
}
\value{
a dataframe with (ndraw - nomit) row, each row is a MCMC draw;
}
\description{
basic version of probit model fitting function.
}
\examples{
#uncomment code to run test
#dt = sim_linear(N=100,beta=c(0.2,-0.8));
#md = linear_fit(y=dt$Y,X=dt$X,n = dt$N,ndraw=10000)
#round(apply(md,2,mean),3);round(apply(md,2,mean)/apply(md,2,sd),3)
#result
#		       mean	   	sd     	true value
#beta_1  0.203   1.806		         0.2
#beta_2 -0.805  -7.702            -0.8
}
