% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_linear.R
\name{panel_linear}
\alias{panel_linear}
\title{function for estimate panel linear model}
\usage{
panel_linear(y, X, n, T, ndraw = 2000, nomit = 100)
}
\arguments{
\item{y}{y is the response;}

\item{X}{X is data frame ccontains data;}

\item{n}{n is the observation number;}

\item{T}{T is the time period of data;}

\item{ndraw}{ndraw is the draw times for MCMC;}

\item{nomit}{noimt is the draw times omit for burn in;}
}
\value{
a dataframe with (ndraw - nomit) row, each row is a MCMC draw;
}
\description{
panel linear model fitting function.
}
\examples{
#uncomment code to run test
#dt = sim_panel_linear(N=100,T=10,beta=c(0.7,-0.5,0.3),alpha=0.5);
#md = panel_linear(y=dt$Y,X=dt$X,n=dt$N,T=dt$T,ndraw=1000)
#round(apply(md,2,mean),3);round(apply(md,2,mean)/apply(md,2,sd),3)
#result
#       beta_1  beta_2  beta_3      mu  xi_sqr
#[1]   0.705  -0.489   0.305   0.598   0.738
#[1]  22.191 -16.025   9.489   6.475   6.086
}
