% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_probit.R
\name{panel_probit}
\alias{panel_probit}
\title{fitting panel probit model}
\usage{
panel_probit(y, X, n, T, ndraw = 2000, nomit = 100)
}
\arguments{
\item{y}{y is binary response;}

\item{X}{X is data frame ccontains data;}

\item{n}{n is the observation number;}

\item{T}{T is the time period of data;}

\item{ndraw}{ndraw is the draw times for MCMC;}

\item{nomit}{noimt is the draw times omit for burn in;}
}
\value{
a dataframe with (ndraw - nomit) row, each row is a MCMC draw;
}
\description{
A function for estimate panel probit model
}
\examples{
#uncomment code to run test
#dt=sim_panel_probit(N=100,T=10,beta=c(0.7,-0.5,0.3),alpha=0.5);
#md=panel_probit(y=dt$Y,X=dt$X,n=dt$N,T=dt$T,ndraw=1000)
#round(apply(md,2,mean),3);round(apply(md,2,mean)/apply(md,2,sd),3)
#result
# 	    beta_1   beta_2  beta_3     mu  xi_sqr
#[1]    0.724   -0.573   0.308  0.593   1.142
#[1]   11.003   -9.525   5.246  4.824   4.195
}
