% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit.R
\name{probit}
\alias{probit}
\title{fitting probit model}
\usage{
probit(y, X, n, ndraw = 2000, nomit = 1000)
}
\arguments{
\item{y}{y is binary response;}

\item{X}{X is data frame ccontains data;}

\item{n}{n is the observation number;}

\item{ndraw}{ndraw is the draw times for MCMC;}

\item{nomit}{noimt is the draw times omit for burn in;}
}
\value{
a dataframe with (ndraw - nomit) row, each row is a MCMC draw;
}
\description{
A function for estimate probit model
}
\examples{
#uncomment code to run test
#library(tmvtnorm)
#dt=sim_probit(N=100,beta=c(0.5,-0.7));
#md=probit(y=dt$Y,X=dt$X,n=dt$N,ndraw=3000)
#round(apply(md,2,mean),3);round(apply(md,2,mean)/apply(md,2,sd),3)
#  result
#  [1]  0.394 -0.685
#  [1]  2.430 -4.183
}
