% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAR.R
\name{SAR}
\alias{SAR}
\title{fitting spatial autoregressive model}
\usage{
SAR(y, X, W, n, ndraw = 2000, nomit = 100)
}
\arguments{
\item{y}{y is the response;}

\item{X}{X is data frame ccontains data;}

\item{W}{W is the spatial weight matrix;}

\item{n}{n is the observation number;}

\item{ndraw}{ndraw is the draw times for MCMC;}

\item{nomit}{noimt is the draw times omit for burn in;}
}
\value{
a dataframe with (ndraw - nomit) row, each row is a MCMC draw;
}
\description{
A function for estimate spatial autoregressive model
}
\examples{
#uncomment code to run test
#dt = sim_SAR(N=100,beta=c(0.5,-0.8),rho=-0.5);
#md = SAR(y=dt$Y,X=apply(dt$X,2,scale),W=dt$W,n=dt$N,ndraw=2000)
#round(apply(md,2,mean),3);round(apply(md,2,mean)/apply(md,2,sd),3)
#Result
#[1]   0.479   -0.825  -0.402
#[1]  22.391  -58.843  -1.800
}
